requirements.txt
python-telegram-bot
openai
python-dotenv
sqlalchemy
numpy
alembic
pandas
